feat: allow lbry uris
lbry.tv has experimental feed support via
[`melroy/lbry-channel-feed`](https://gitlab.melroy.org/melroy/lbry-channel-feed).
Unfortunatly it was unusable using nextcloud news because the `lbry`
scheme wasn't whitelisted.

To fix that, extend the default list of schemes, htmlpurifier allows
with the `lbry` scheme.

For future work it might make sense to review the defaults and maybe
remove some of them, since not all look like they make sense for a
feedreader, but for backwards compatability and to keep this change
focused I left them untouched for now.